update: 更新

+++ src/pages/retro/section/index.tsx
@@ -132,8 +132,11 @@ const Section = () => {
   };
 
   const renderItem = (i) => {
+    const hasAuth = isCreator || i.user._id === userId;
     return (
       <Item
+        isCreator={isCreator}
+        hasAuth={hasAuth}
         anonymous={data?.retro?.anonymous || i.anonymous}
         blur={hasFocus && i.status !== 'FOCUSED'}
         key={i._id}
@@ -144,12 +147,18 @@ const Section = () => {
         type={i.type}
         like={i.like}
         onDelete={() => {
+          if (!isCreator) {
+            return;
+          }
           handleDelete(i._id);
         }}
         onLike={(count: number) => {
           handleLike(i._id, count);
         }}
         onUpdateContent={(variables: UpdateParams) => {
+          if (!isCreator) {
+            return;
+          }
           handleUpdate(i._id, variables);
         }}
         onUpdateStatus={(variables: UpdateParams) => {
